requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
reportlab==4.0.4
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
